[tests] Merge abstract and sub type class
